refactor: extract useServerEvent
